Move Java-related AST extension tests from `:compiler` to `:codegen-java`
